Address PR comments: better validation, reorganize tests
